requirements.txt
fastapi==0.78.0
numpy==1.19.5
Pillow==9.2.0
tensorflow==2.5.0
uvicorn==0.17.6
aiofiles==0.6.0